Switch to using `eslint` package directly instead of `eslint-linter-browserify`
