+++ crates/brioche/runtime/package.json
@@ -14,13 +14,15 @@
   "dependencies": {
     "@typescript-eslint/parser": "^6.19.0",
     "buffer": "^6.0.3",
-    "eslint-linter-browserify": "^8.56.0",
+    "crypto-browserify": "^3.12.0",
+    "eslint": "^8.56.0",
     "events": "^3.3.0",
     "os-browserify": "^0.3.0",
     "path-browserify": "^1.0.1",
     "stream-browserify": "^3.0.0",
     "ts-node": "^10.9.2",
     "typescript": "^5.3.3",
+    "url": "^0.11.3",
     "vscode-languageserver": "^8.1.0"
   }
 }
