+++ crates/brioche/runtime/src/compat/node/assert.cjs
@@ -1,0 +1,1 @@
+module.exports = () => {};
