+++ crates/brioche/runtime/src/compat/node/buffer.cjs
@@ -1,1 +1,1 @@
-module.exports = require("buffer");
+module.exports = require("buffer/");
