+++ crates/brioche/runtime/src/compat/node/crypto.cjs
@@ -1,0 +1,1 @@
+module.exports = require("crypto-browserify");
