+++ crates/brioche/runtime/src/compat/node/events.cjs
@@ -1,0 +1,1 @@
+module.exports = require("events/");
