+++ crates/brioche/runtime/src/compat/node/module.cjs
@@ -1,0 +1,1 @@
+module.exports.createRequire = () => {};
