+++ crates/brioche/runtime/src/compat/node/url.cjs
@@ -1,0 +1,2 @@
+module.exports = require("url");
+module.exports.pathToFileURL = (path) => `file://${path}`;
