+++ crates/brioche/runtime/src/index.ts
@@ -1,4 +1,4 @@
-import "./text-encoding";
+import "./compat/setup.cjs";
 
 export { check } from "./check.ts";
 export { buildLsp } from "./lsp.ts";
